Moved functions/components related to overview into overview
